pyproject.toml
[project]
name = "proxyshop"
version = "1.13.2"
description = "Photoshop automation tool for generating high quality Magic the Gathering card renders."
authors = [{name = "Investigamer", email = "freethoughtleft@gmail.com"}]
license = "MPL-2.0"
readme = "README.md"
keywords = [
    "proxyshop",
    "proxy",
    "mtg",
    "magic",
    "gathering",
    "cards",
    "photoshop",
    "magic the gathering",
    "playtest",
]
requires-python = ">=3.14,<3.15"

[tool.poetry.urls]
Changelog = "https://github.com/Investigamer/Proxyshop/blob/main/CHANGELOG.md"
Discord = "https://discord.gg/magicproxies"
Documentation = "https://investigamer.github.io/Proxyshop"
Issues = "https://github.com/Investigamer/Proxyshop/issues"
Source = "https://github.com/Investigamer/Proxyshop"
Sponsor = "https://patreon.com/mpcfill"

[[tool.poetry.packages]]
include = 'src/../'

[tool.poetry.dependencies]
photoshop-python-api = {git = "https://github.com/pappnu/photoshop-python-api.git", branch = "type-annotation"}
requests = "^2.32.5"
Pillow = "^12.1.1"
typing-extensions = "^4.15.0"
limits = "^5.8.0"
backoff = "^2.2.1"
pathvalidate = "^3.3.1"
fonttools = "^4.61.1"
pyyaml = "^6.0.3"
tqdm = "^4.67.3"
click = "^8.3.1"
tomli = "^2.4.0"
yarl = "^1.22.0"
pydantic = "^2.12.5"
pydantic-settings = "^2.13.0"
omnitils = {git = "https://github.com/pappnu/omnitils.git", branch = "dev"}
hexproof = {git = "https://github.com/pappnu/hexproof.git", branch = "dev"}
rich = "^14.3.2"
pywin32 = "^311.0.0"
pyside6 = "^6.10.2"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
mypy = "^1.19.1"
commitizen = "^4.13.7"
setuptools = "^82.0.0"
matplotlib = "^3.10.8"
psd-tools = "^1.12.1"
pyclean = "^3.5.0"
pyinstaller = "^6.19.0"
pre-commit = "^4.5.1"
mkdocs = "^1.6.1"
mkdocs-material = "^9.7.1"
mkdocs-include-markdown-plugin = "^7.2.1"
mkdocs-pymdownx-material-extras = "^2.8.0"
mkdocs-minify-plugin = "^0.8.0"
mkdocstrings-python = "^2.0.2"
mkdocs-gen-files = "^0.6.0"
mkdocs-autolinks-plugin = "^0.7.1"
mkdocs-same-dir = "^0.1.3"
mkdocs-git-revision-date-plugin = "^0.3.2"
mkdocstrings = {extras = ["python"], version = "^1.0.3"}
memory-profiler = "^0.61.0"
types-pywin32 = "^311.0.0.20251008"
types-requests = "^2.32.4.20260107"

[tool.poetry.scripts]
proxyshop = 'main:launch_cli'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.commitizen]
version_provider = "pep621"
encoding = "utf-8"
changelog_start_rev = 'v1.2.0'
tag_format = "v$major.$minor.$patch"
update_changelog_on_bump = true

[tool.ruff.lint]
extend-select = [
    # isort
    "I",
    # pyupgrade
    "UP",
]
ignore = ["F403", "UP015"]

[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = "information"
